requirements.txt
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
boto3==1.7.50
botocore==1.10.50
certifi==2018.4.16
chardet==3.0.4
coloredlogs==10.0
coverage==4.5.1
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
humanfriendly==4.17
idna==2.7
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.2.0
pluggy==0.6.0
prompt-toolkit==2.0.7
py==1.5.4
pyaml==17.12.1
pylint==2.1.1
pytest==3.6.3
pytest-cov==2.5.1
python-dateutil==2.7.3
python-dotenv==0.9.1
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
tabulate==0.8.2
urllib3==1.23
wcwidth==0.1.7
wrapt==1.10.11
